+++ src/App.js
@@ -1,5 +1,6 @@
 import { Route, Routes } from 'react-router-dom';
 import { ROUTE } from './common/utils/constant';
+import Header from './components/Header/Header';
 import Main from './pages/Main/Main';
 import MovieDetail from './pages/MovieDetail/MovieDetail';
 import NotFound from './pages/NotFound/NotFound';
@@ -7,12 +8,15 @@ import Search from './pages/Search/Search';
 
 function App() {
   return (
-    <Routes>
-      <Route path={ROUTE.MAIN} element={<Main />} />
-      <Route path={ROUTE.NOT_FOUND} element={<NotFound />} />
-      <Route path={ROUTE.MOVIE_ID} element={<MovieDetail />} />
-      <Route path={ROUTE.SEARCH} element={<Search />} />
-    </Routes>
+    <>
+      <Header />
+      <Routes>
+        <Route path={ROUTE.MAIN} element={<Main />} />
+        <Route path={ROUTE.NOT_FOUND} element={<NotFound />} />
+        <Route path={ROUTE.MOVIE_ID} element={<MovieDetail />} />
+        <Route path={ROUTE.SEARCH} element={<Search />} />
+      </Routes>
+    </>
   );
 }
